+++ src/constants/auth.ts
@@ -1,0 +1,6 @@
+/**
+ * Used to obtain public key on server startup, so user token on login can be verified.
+ */
+export const AUTH_LOGIN_SERVER_KEY_URL = 'https://login.airmash.online/key';
+
+export const AUTH_LOGIN_SERVER_DOWNLOAD_RETRY_INTERVAL_SEC = 5;