+++ src/constants/index.ts
@@ -17,3 +17,5 @@ export * from '@/constants/mountains';
 export * from '@/constants/ffa';
 export * from '@/constants/ctf';
 export * from '@/constants/btr';
+export * from '@/constants/user-stats';
+export * from '@/constants/auth';