+++ src/server/modes/base/controllers/login.ts
@@ -61,13 +61,41 @@ export default class LoginMessageHandler extends System {
     /**
      * Validation
      */
-    const user = false;
+    let userId: string;
     let { flag, name } = msg;
 
     if (msg.protocol !== 5) {
       this.emit(ERRORS_INCORRECT_PROTOCOL, connectionId);
 
       return;
+    }
+
+    if (msg.session !== 'none') {
+      let validSessionData = true;
+
+      try {
+        const session = JSON.parse(msg.session);
+
+        if (session.token) {
+          if (typeof session.token === 'string') {
+            userId = this.helpers.getUserIdFromToken(session.token);
+
+            if (!userId) {
+              validSessionData = false;
+            }
+          } else {
+            validSessionData = false;
+          }
+        }
+      } catch (e) {
+        validSessionData = false;
+      }
+
+      if (!validSessionData) {
+        this.emit(ERRORS_INVALID_LOGIN_DATA, connectionId);
+
+        return;
+      }
     }
 
     if (this.app.config.allowNonAsciiUsernames === false) {
@@ -106,7 +134,7 @@ export default class LoginMessageHandler extends System {
         x: msg.horizonX,
         y: msg.horizonY,
       },
-      user,
+      userId,
       shipType: SHIPS_TYPES.PREDATOR,
     });
   }