+++ src/server/modes/base/responses/score-update.ts
@@ -1,5 +1,5 @@
-import { SERVER_PACKETS, ServerPackets } from '@airbattle/protocol';
-import { RESPONSE_SCORE_UPDATE, CONNECTIONS_SEND_PACKET } from '@/events';
+import { PLAYER_LEVEL_UPDATE_TYPES, SERVER_PACKETS, ServerPackets } from '@airbattle/protocol';
+import { RESPONSE_PLAYER_LEVEL, RESPONSE_SCORE_UPDATE, CONNECTIONS_SEND_PACKET } from '@/events';
 import { System } from '@/server/system';
 import { PlayerId } from '@/types';
 
@@ -26,16 +26,35 @@ export default class ScoreUpdate extends System {
     const player = this.storage.playerList.get(playerId);
     const connectionId = this.storage.playerMainConnectionList.get(playerId);
 
+    let earnings = 0;
+    let totalkills = player.kills.current;
+    let totaldeaths = player.deaths.current;
+
+    if (player.user) {
+      const user = this.storage.userList.get(player.user.id);
+
+      earnings = user.lifetimestats.earnings;
+      totalkills = user.lifetimestats.totalkills;
+      totaldeaths = user.lifetimestats.totaldeaths;
+
+      const newLevel = this.helpers.convertEarningsToLevel(earnings);
+
+      if (newLevel > player.level.current) {
+        player.level.current = newLevel;
+        this.emit(RESPONSE_PLAYER_LEVEL, player.id.current, PLAYER_LEVEL_UPDATE_TYPES.LEVELUP);
+      }
+    }
+
     this.emit(
       CONNECTIONS_SEND_PACKET,
       {
         c: SERVER_PACKETS.SCORE_UPDATE,
         id: player.id.current,
         score: player.score.current,
-        earnings: player.earningscore.current,
+        earnings,
         upgrades: player.upgrades.amount,
-        totalkills: player.kills.total,
-        totaldeaths: player.deaths.total,
+        totalkills,
+        totaldeaths,
       } as ServerPackets.ScoreUpdate,
       connectionId
     );